Implement virtual functions
- Keep track of which functions are actually called
- Keep track of which functions are overridden and by which class
- Always type and generate functions that override others, even if they
are not called. (They might be indirectly called and the type inferrer
might not catch that)
- Add a new int32 (…)** in front of the structure instead of the
previous vtable pointer responsible for storing the vtable.
- Store pointers to methods in the vtable and do late binding instead
of early binding

All of this seems to be working, but I have not tested every single
possibility. I also need to check if the late binding can be optimized
away in some situations.